# backend/openapi/TidySense.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "TidySense | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/reconcile/prompt": {
      "post": {
        "tags": [
          "Reconcile"
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveReconcilePromptRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveReconcilePromptRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveReconcilePromptRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReconcilePromptDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReconcilePromptDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReconcilePromptDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reconcile/sessions": {
      "post": {
        "tags": [
          "Reconcile"
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenReconcileSessionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenReconcileSessionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/OpenReconcileSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReconcileSessionDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReconcileSessionDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReconcileSessionDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reconcile/sessions/{id}": {
      "get": {
        "tags": [
          "Reconcile"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReconcileSessionDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReconcileSessionDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReconcileSessionDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reconcile/sessions/{id}/complete": {
      "post": {
        "tags": [
          "Reconcile"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteReconcileSessionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteReconcileSessionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteReconcileSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReconcileSessionDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReconcileSessionDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReconcileSessionDto"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reconcile/sessions/{id}/previews": {
      "post": {
        "tags": [
          "Reconcile"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReconcilePreviewRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReconcilePreviewRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReconcilePreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ActionConfirmationDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionConfirmationDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionConfirmationDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reconcile/confirmations/{id}/submit": {
      "post": {
        "tags": [
          "Reconcile"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitConfirmationRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitConfirmationRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitConfirmationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ConfirmationResultDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfirmationResultDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfirmationResultDto"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AcknowledgedWarningDto": {
        "required": [
          "warningId",
          "warningHash"
        ],
        "type": "object",
        "properties": {
          "warningId": {
            "type": "string"
          },
          "warningHash": {
            "type": "string"
          }
        }
      },
      "ActionConfirmationDto": {
        "required": [
          "id",
          "reconcileSessionId",
          "actionType",
          "status",
          "canApply",
          "items",
          "warnings",
          "previewHash",
          "expiresAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "reconcileSessionId": {
            "type": "string",
            "format": "uuid"
          },
          "actionType": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "canApply": {
            "type": "boolean"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReconcilePreviewItemDto"
            }
          },
          "warnings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConfirmationWarningDto"
            }
          },
          "previewHash": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ApiProblemDto": {
        "required": [
          "type",
          "title",
          "status",
          "code",
          "messageKey",
          "retryable",
          "traceId",
          "correlationId",
          "detail",
          "errors",
          "entityId",
          "expectedVersion",
          "currentVersion"
        ],
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "code": {
            "type": "string"
          },
          "messageKey": {
            "type": "string"
          },
          "retryable": {
            "type": "boolean"
          },
          "traceId": {
            "type": "string"
          },
          "correlationId": {
            "type": "string"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": [
              "null",
              "object"
            ],
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          "entityId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "expectedVersion": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "currentVersion": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int64"
          }
        }
      },
      "CaptureDto": {
        "required": [
          "id",
          "title",
          "status",
          "source",
          "version",
          "createdAt",
          "updatedAt",
          "resolvedAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "version": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "resolvedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
...
      "CompleteReconcileSessionRequest": {
        "required": [
          "expectedVersion"
        ],
        "type": "object",
        "properties": {
          "expectedVersion": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          }
        }
      },
...
      "ConfirmationResultDto": {
        "required": [
          "confirmationId",
          "status",
          "actionType",
          "affectedCount"
        ],
        "type": "object",
        "properties": {
          "confirmationId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string"
          },
          "actionType": {
            "type": "string"
          },
          "affectedCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
      "CreateReconcilePreviewRequest": {
        "required": [
          "actionType",
          "taskIds",
          "sequenceId",
          "plannedDate",
          "includeTaskIds"
        ],
        "type": "object",
        "properties": {
          "actionType": {
            "type": "string"
          },
          "taskIds": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "sequenceId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "plannedDate": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "includeTaskIds": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        }
      },
...
      "OpenReconcileSessionRequest": {
        "required": [
          "triggerType"
        ],
        "type": "object",
        "properties": {
          "triggerType": {
            "type": "string"
          }
        }
      },
...
      "ReconcileCountsDto": {
        "required": [
          "actionableBacklogCount",
          "oldestUnresolvedAgeDays",
          "repeatedCarryTaskCount",
          "deadlineRiskCount",
          "affectedParentCount",
          "reviewDueCount",
          "unresolvedCaptureCount"
        ],
        "type": "object",
        "properties": {
          "actionableBacklogCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "oldestUnresolvedAgeDays": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "repeatedCarryTaskCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "deadlineRiskCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "affectedParentCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "reviewDueCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "unresolvedCaptureCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
      "ReconcileOwnerGroupDto": {
        "required": [
          "ownerType",
          "ownerId",
          "ownerTitle",
          "sequences",
          "tasks"
        ],
        "type": "object",
        "properties": {
          "ownerType": {
            "type": "string"
          },
          "ownerId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "ownerTitle": {
            "type": [
              "null",
              "string"
            ]
          },
          "sequences": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReconcileSequenceGroupDto"
            }
          },
          "tasks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReconcileTaskItemDto"
            }
          }
        }
      },
      "ReconcilePreviewItemDto": {
        "required": [
          "taskId",
          "title",
          "expectedVersion",
          "classification",
          "currentPlannedDate",
          "resultingPlannedDate",
          "resultingStatus"
        ],
        "type": "object",
        "properties": {
          "taskId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "expectedVersion": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "classification": {
            "type": "string"
          },
          "currentPlannedDate": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "resultingPlannedDate": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "resultingStatus": {
            "type": "string"
          }
        }
      },
      "ReconcilePromptDto": {
        "required": [
          "localDate",
          "state"
        ],
        "type": "object",
        "properties": {
          "localDate": {
            "type": "string",
            "format": "date"
          },
          "state": {
            "type": "string"
          }
        }
      },
      "ReconcileReviewItemDto": {
        "required": [
          "entityType",
          "id",
          "title",
          "version",
          "reviewDate",
          "targetDate",
          "allowedActions",
          "undatedTasks"
        ],
        "type": "object",
        "properties": {
          "entityType": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "version": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "reviewDate": {
            "type": "string",
            "format": "date"
          },
          "targetDate": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "allowedActions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "undatedTasks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReconcileTaskRefDto"
            }
          }
        }
      },
      "ReconcileRuleMatchDto": {
        "required": [
          "ruleId",
          "ruleVersion",
          "affectedEntityIds",
          "allowedActions"
        ],
        "type": "object",
        "properties": {
          "ruleId": {
            "type": "string"
          },
          "ruleVersion": {
            "type": "string"
          },
          "affectedEntityIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "allowedActions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "ReconcileSessionDto": {
        "required": [
          "id",
          "status",
          "version",
          "localDate",
          "rulesCatalogVersion",
          "openedAt",
          "completedAt",
          "openedSeverity",
          "openedActionableBacklogCount",
          "severity",
          "triggerReasons",
          "counts",
          "executionGroups",
          "commitmentReviews",
          "captures",
          "ruleMatches"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string"
          },
          "version": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "localDate": {
            "type": "string",
            "format": "date"
          },
          "rulesCatalogVersion": {
            "type": "string"
          },
          "openedAt": {
            "type": "string",
            "format": "date-time"
          },
          "completedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "openedSeverity": {
            "type": "string"
          },
          "openedActionableBacklogCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "severity": {
            "type": "string"
          },
          "triggerReasons": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "counts": {
            "$ref": "#/components/schemas/ReconcileCountsDto"
          },
          "executionGroups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReconcileOwnerGroupDto"
            }
          },
          "commitmentReviews": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReconcileReviewItemDto"
            }
          },
          "captures": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CaptureDto"
            }
          },
          "ruleMatches": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReconcileRuleMatchDto"
            }
          }
        }
      },
...
      "ResolveReconcilePromptRequest": {
        "required": [
          "state"
        ],
        "type": "object",
        "properties": {
          "state": {
            "type": "string"
          }
        }
      },
...
      "SubmitConfirmationRequest": {
        "required": [
          "acknowledgedWarnings"
        ],
        "type": "object",
        "properties": {
          "acknowledgedWarnings": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/AcknowledgedWarningDto"
            }
          }
        }
      },
...
    }
  },
...
}